deploy/requirements.txt
flask
psutil
gevent
cloudant
confluent_kafka
PyYAML
dill
pika==0.13.1
python-dateutil
docker
boto3
jsonpath_ng
redis